refactor: lib structure
