+++ buncha/package.json
@@ -25,6 +25,7 @@
   "license": "ISC",
   "dependencies": {
     "@prisma/client": "^4.4.0",
+    "axios": "^1.1.3",
     "cors": "^2.8.5",
     "dotenv": "^16.0.3",
     "express": "^4.18.1",
