+++ buncha/src/index.ts
@@ -1,4 +1,4 @@
+import { logger } from '@/lib/utils/logger';
 import main from '@/main';
-import { logger } from '@/utils/logger';
 
 main().catch((err) => logger.error(err));
