+++ buncha/src/lib/apis/space.ts
@@ -1,0 +1,3 @@
+export const getAllSpaces = async () => {
+  //
+};
