+++ buncha/src/lib/types/ws.ts
